backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
motor==3.3.2  # MongoDB async driver
python-dotenv==1.0.0
openai
supabase
